# src/doc/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"securitySchemes": {}
	},
	"info": {
		"title": "back",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/user/program/{id}": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"userId": {
											"type": "number",
											"format": "double"
										},
										"fileUri": {
											"type": "string"
										},
										"id": {
											"type": "number",
											"format": "double"
										},
										"name": {
											"type": "string"
										}
									},
									"required": [
										"userId",
										"fileUri",
										"id",
										"name"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Programmes"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"programFile": {
										"type": "string",
										"format": "binary"
									}
								},
								"required": [
									"programFile"
								]
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"userId": {
											"type": "number",
											"format": "double"
										},
										"fileUri": {
											"type": "string"
										},
										"id": {
											"type": "number",
											"format": "double"
										},
										"name": {
											"type": "string"
										}
									},
									"required": [
										"userId",
										"fileUri",
										"id",
										"name"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Programmes"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"name": {
										"type": "string"
									}
								},
								"required": [
									"name"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}